Remove analysis dir
