+++ CMakeLists.txt
@@ -177,6 +177,7 @@ file(GLOB TILE_LANG_SRCS
   src/layout/*.cc
   src/tileview/*.cc
   src/transform/*.cc
+  src/transform/sunmmio_tile_loop_fusion/*.cc
   src/transform/common/*.cc
   src/op/*.cc
   src/target/utils.cc
