+++ src/transform/sunmmio_tile_loop_fusion/utils.h
@@ -1,12 +1,16 @@
+/*!
+ * \file utils.h
+ * \brief Shared logical-axis normalization and small helper utilities for
+ * Sunmmio tile loop fusion.
+ *
+ * Discovery and planning both need the same canonicalization machinery for
+ * execution-axis names, normalized buffer regions, and lightweight expression
+ * inspection. This header keeps those helpers in one place without introducing
+ * an additional stage boundary of its own.
+ */
 #pragma once
 
-// Utility helpers shared by discovery and planning.
-//
-// These helpers do not define stage boundaries themselves. They exist to keep
-// recurring logical-axis normalization and variable-collection code out of the
-// higher-level analysis and planner entrypoints.
-
-#include "sunmmio_tile_loop_fusion_protocol.h"
+#include "types.h"
 
 #include <tvm/tir/stmt_functor.h>
 
@@ -18,10 +22,18 @@
 namespace tvm {
 namespace tl {
 
+/*! \brief Render a PrimExpr into a stable string for diagnostics and keys. */
 String PrimExprToString(const PrimExpr &expr);
 
+/*! \brief Return a printable name for one dependence kind. */
 const char *DependenceKindToCString(TileScopeDependenceKind kind);
 
+/*!
+ * \brief Collect every Var referenced by the visited expression subtree.
+ *
+ * Discovery uses this to determine which normalized access dimensions depend on
+ * which execution axes.
+ */
 class VarUseCollector : public tir::ExprVisitor {
 public:
   std::unordered_set<const tir::VarNode *> seen_vars;
@@ -30,10 +42,18 @@ private:
   void VisitExpr_(const tir::VarNode *op) final;
 };
 
+/*!
+ * \brief Replace a region's lowered execution-loop vars with canonical logical
+ * axis vars such as `i`, `j`, and `k`.
+ */
 Map<tir::Var, PrimExpr> BuildLogicalExecutionAxisSubstitution(
     const TileScopeRegion &region,
     std::unordered_map<std::string, tir::Var> *canonical_execution_vars);
 
+/*!
+ * \brief Rewrite one BufferRegion into the shared logical-axis coordinate
+ * space produced by \ref BuildLogicalExecutionAxisSubstitution.
+ */
 tir::BufferRegion NormalizeBufferRegionByLogicalExecutionAxes(
     const tir::BufferRegion &region, const Map<tir::Var, PrimExpr> &subst);
